Move aspect specific classes to aspect package
