+++ core/esmf-turtle-language-server/src/main/java/org/eclipse/esmf/turtle/languageserver/aspect/service/AspectModelValidationService.java
@@ -25,17 +25,13 @@ import org.eclipse.esmf.aspectmodel.shacl.violation.Violation;
 import org.eclipse.esmf.aspectmodel.validation.ProcessingViolation;
 import org.eclipse.esmf.aspectmodel.validation.Validator;
 import org.eclipse.esmf.aspectmodel.validation.services.AspectModelValidator;
-import org.eclipse.esmf.metamodel.vocabulary.RdfNamespace;
-import org.eclipse.esmf.metamodel.vocabulary.SammNs;
-import org.eclipse.esmf.treesitterturtle.ParserTokenType;
-import org.eclipse.esmf.treesitterturtle.TurtleSyntaxTree;
 import org.eclipse.esmf.turtle.languageserver.aspect.diagnostic.AspectViolationDiagnosticMapper;
+import org.eclipse.esmf.turtle.languageserver.aspect.navigation.ExternalModelFileCache;
 import org.eclipse.esmf.turtle.languageserver.lsp.ResolutionStrategyService;
 import org.eclipse.esmf.turtle.languageserver.lsp.diagnostic.DiagnosticReport;
 import org.eclipse.esmf.turtle.languageserver.lsp.diagnostic.ResolutionStrategyAwareDiagnosticsProvider;
 import org.eclipse.esmf.turtle.languageserver.lsp.text.ParsedDocument;
 import org.eclipse.esmf.turtle.languageserver.turtle.TurtleService;
-import org.eclipse.esmf.turtle.languageserver.turtle.navigation.ExternalModelFileCache;
 
 import org.apache.jena.riot.RiotException;
 import org.slf4j.Logger;
@@ -43,7 +39,6 @@ import org.slf4j.LoggerFactory;
 
 public class AspectModelValidationService extends TurtleService implements ResolutionStrategyAwareDiagnosticsProvider {
    private static final Logger LOG = LoggerFactory.getLogger( AspectModelValidationService.class );
-   private static final List<String> SAMM_PREFIXES = SammNs.sammNamespaces().map( RdfNamespace::getShortForm ).toList();
    private final Validator<Violation, List<Violation>> validator;
    private final AspectViolationDiagnosticMapper diagnosticMapper = new AspectViolationDiagnosticMapper();
    private ResolutionStrategyService resolutionStrategyService;
@@ -60,16 +55,6 @@ public class AspectModelValidationService extends TurtleService implements Resol
          final ResolutionStrategyService resolutionStrategyService ) {
       this.validator = validator;
       this.resolutionStrategyService = resolutionStrategyService;
-   }
-
-   private boolean documentIsAspectModel( final ParsedDocument parsedDocument ) {
-      final TurtleSyntaxTree syntaxTree = parsedDocument.turtleSyntaxTree();
-      return syntaxTree.rootNode().children().stream()
-            .filter( n -> ParserTokenType.DIRECTIVE.equals( n.type() ) ).flatMap( n -> n.children().stream() )
-            .filter( n -> ParserTokenType.PREFIX_ID.equals( n.type() ) ).flatMap( n -> n.children().stream() )
-            .filter( n -> ParserTokenType.NAMESPACE.equals( n.type() ) ).flatMap( n -> n.children().stream() )
-            .filter( n -> ParserTokenType.PN_PREFIX.equals( n.type() ) )
-            .anyMatch( n -> SAMM_PREFIXES.contains( n.content() ) );
    }
 
    private boolean shouldValidateDocument( final ParsedDocument parsedDocument ) {
