+++ core/esmf-turtle-language-server/src/main/java/org/eclipse/esmf/turtle/languageserver/lsp/text/TurtleTextDocumentService.java
@@ -31,7 +31,7 @@ import org.eclipse.esmf.turtle.languageserver.structure.DocumentSymbolService;
 import org.eclipse.esmf.turtle.languageserver.structure.TurtleTokenService;
 import org.eclipse.esmf.turtle.languageserver.turtle.TurtleCompletionService;
 import org.eclipse.esmf.turtle.languageserver.turtle.ValidationCoordinator;
-import org.eclipse.esmf.turtle.languageserver.turtle.navigation.TurtleCrossFileDefinitionService;
+import org.eclipse.esmf.turtle.languageserver.aspect.navigation.AspectCrossFileDefinitionService;
 import org.eclipse.esmf.turtle.languageserver.turtle.navigation.TurtleDefinitionService;
 
 import org.eclipse.lsp4j.CompletionItem;
@@ -63,7 +63,7 @@ public class TurtleTextDocumentService implements TextDocumentService {
 
    private final TextDocumentClientNotifier clientNotifier;
    private final TurtleDefinitionService turtleDefinitionService;
-   private final TurtleCrossFileDefinitionService turtleCrossFileDefinitionService;
+   private final AspectCrossFileDefinitionService aspectCrossFileDefinitionService;
    private final TurtleCompletionService turtleCompletionService;
    private final ValidationCoordinator validationCoordinator;
    private final TreeSitterTurtleParserService turtleParserService;
@@ -84,7 +84,7 @@ public class TurtleTextDocumentService implements TextDocumentService {
       turtleCompletionService = new TurtleCompletionService();
       turtleParserService = new TreeSitterTurtleParserService();
       tokenService = new TurtleTokenService( turtleParserService );
-      turtleCrossFileDefinitionService = new TurtleCrossFileDefinitionService( turtleParserService, documents, resolutionStrategyService );
+      aspectCrossFileDefinitionService = new AspectCrossFileDefinitionService( turtleParserService, documents, resolutionStrategyService );
       documentSymbolService = new DocumentSymbolService( turtleParserService );
       final List<DiagnosticsProvider> diagnosticsProviders =
             Streams.stream( ServiceLoader.load( DiagnosticsProvider.class ).iterator() ).toList();
@@ -195,8 +195,8 @@ public class TurtleTextDocumentService implements TextDocumentService {
          final ParsedDocument parsedDocument = turtleParserService.apply( document );
          Optional<Location> declaration = turtleDefinitionService.findDefinition( parsedDocument, params.getPosition() );
 
-         if ( declaration.isEmpty() && turtleCrossFileDefinitionService != null ) {
-            declaration = turtleCrossFileDefinitionService.findDefinition( parsedDocument, params.getPosition() );
+         if ( declaration.isEmpty() && aspectCrossFileDefinitionService != null ) {
+            declaration = aspectCrossFileDefinitionService.findDefinition( parsedDocument, params.getPosition() );
          }
 
          return declaration
