+++ core/esmf-turtle-language-server/src/main/java/org/eclipse/esmf/turtle/languageserver/turtle/TurtleService.java
@@ -19,6 +19,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
 
+import org.eclipse.esmf.metamodel.vocabulary.RdfNamespace;
+import org.eclipse.esmf.metamodel.vocabulary.SammNs;
 import org.eclipse.esmf.treesitterturtle.ParserTokenType;
 import org.eclipse.esmf.treesitterturtle.TurtleSyntaxTree;
 import org.eclipse.esmf.turtle.languageserver.lsp.text.ParsedDocument;
@@ -31,6 +33,7 @@ import org.eclipse.lsp4j.Range;
  * Base class for services that operate on a {@link TurtleSyntaxTree}
  */
 public abstract class TurtleService {
+   private static final List<String> SAMM_PREFIXES = SammNs.sammNamespaces().map( RdfNamespace::getShortForm ).toList();
    protected static final List<String> TYPE_DEFINITION_PREDICATES = List.of( "a", "rdf:type" );
 
    /**
@@ -43,6 +46,16 @@ public abstract class TurtleService {
       final Map<TurtleSyntaxTree.Node, TurtleSyntaxTree.Node> predicateObjectMap = predicateObjectMapForTriple( triple );
       return predicateObjectMap.keySet().stream().map( TurtleSyntaxTree.Node::content )
             .anyMatch( TYPE_DEFINITION_PREDICATES::contains );
+   }
+
+   protected boolean documentIsAspectModel( final ParsedDocument parsedDocument ) {
+      final TurtleSyntaxTree syntaxTree = parsedDocument.turtleSyntaxTree();
+      return syntaxTree.rootNode().children().stream()
+            .filter( n -> ParserTokenType.DIRECTIVE.equals( n.type() ) ).flatMap( n -> n.children().stream() )
+            .filter( n -> ParserTokenType.PREFIX_ID.equals( n.type() ) ).flatMap( n -> n.children().stream() )
+            .filter( n -> ParserTokenType.NAMESPACE.equals( n.type() ) ).flatMap( n -> n.children().stream() )
+            .filter( n -> ParserTokenType.PN_PREFIX.equals( n.type() ) )
+            .anyMatch( n -> SAMM_PREFIXES.contains( n.content() ) );
    }
 
    protected boolean isPrefixedBy( final TurtleSyntaxTree.Node prefixedName, final String prefix ) {
